requirements.txt
numpy
pandas
matplotlib
loguru
h5py
seaborn
scikit-learn
PyYAML
tqdm
librosa
torch
torchlibrosa